[5주차] 심화과제 - 장바구니 유틸 함수 분리
