+++ packages/assignment-5/src/refactoring/pages/cart/hooks/useCart.ts
@@ -1,7 +1,7 @@
-// useCart.ts
 import { useState } from 'react';
 import { CartItem, Coupon, Product } from '../../../../types';
-import { calculateCartTotal, updateCartItemQuantity } from '../utils/cartUtils';
+import { updateCartItemQuantity } from '../utils/cartItemOperations';
+import { calculateCartTotal } from '../utils/cartCalculations';
 
 export const useCart = () => {
   const [cart, setCart] = useState<CartItem[]>([]);
