+++ packages/assignment-5/src/refactoring/pages/cart/utils/cartItemOperations.ts
@@ -1,0 +1,14 @@
+import { CartItem } from '../../../../types';
+
+export const updateCartItemQuantity = (cart: CartItem[], productId: string, newQuantity: number): CartItem[] => {
+  return cart
+    .map((item) => {
+      if (item.product.id === productId) {
+        const maxQuantity = item.product.stock;
+        const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity));
+        return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null;
+      }
+      return item;
+    })
+    .filter((item): item is CartItem => item !== null);
+};
