--- packages/assignment-5/src/refactoring/pages/cart/utils/cartUtils.ts
@@ -1,54 +1,0 @@
-import { CartItem, Coupon } from '../../../../types';
-
-export const calculateItemTotal = (item: CartItem) => {
-  const { price } = item.product;
-  const { quantity } = item;
-
-  const discount = getMaxApplicableDiscount(item);
-
-  return price * quantity * (1 - discount);
-};
-
-export const getMaxApplicableDiscount = (item: CartItem) => {
-  return item.product.discounts.reduce((maxDiscount, d) => {
-    return item.quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount;
-  }, 0);
-};
-
-export const calculateCartTotal = (cart: CartItem[], selectedCoupon: Coupon | null) => {
-  // FIXME: reduce를 두 번 쓰는 게 좋은 걸까?
-  const totalBeforeDiscount = cart.reduce((acc, { product, quantity }) => (acc += product.price * quantity), 0);
-
-  // FIXME: let 안 쓰는 방법이 있을까?..
-  let totalAfterDiscount = cart.reduce((acc, cartItem) => (acc += calculateItemTotal(cartItem)), 0);
-
-  // 쿠폰 적용
-  if (selectedCoupon) {
-    if (selectedCoupon.discountType === 'amount') {
-      totalAfterDiscount = Math.max(0, totalAfterDiscount - selectedCoupon.discountValue);
-    } else {
-      totalAfterDiscount *= 1 - selectedCoupon.discountValue / 100;
-    }
-  }
-
-  const totalDiscount = totalBeforeDiscount - totalAfterDiscount;
-
-  return {
-    totalBeforeDiscount: Math.round(totalBeforeDiscount),
-    totalAfterDiscount: Math.round(totalAfterDiscount),
-    totalDiscount: Math.round(totalDiscount)
-  };
-};
-
-export const updateCartItemQuantity = (cart: CartItem[], productId: string, newQuantity: number): CartItem[] => {
-  return cart
-    .map((item) => {
-      if (item.product.id === productId) {
-        const maxQuantity = item.product.stock;
-        const updatedQuantity = Math.max(0, Math.min(newQuantity, maxQuantity));
-        return updatedQuantity > 0 ? { ...item, quantity: updatedQuantity } : null;
-      }
-      return item;
-    })
-    .filter((item): item is CartItem => item !== null);
-};
