+++ packages/assignment-5/src/refactoring/pages/cart/utils/discountCalculations.ts
@@ -1,0 +1,7 @@
+import { CartItem } from '../../../../types';
+
+export const getMaxApplicableDiscount = (item: CartItem) => {
+  return item.product.discounts.reduce((maxDiscount, d) => {
+    return item.quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount;
+  }, 0);
+};
